Made the tests runnable with jest
Use npm run tutorial to run all the test cases in the tutorial directory.
