+++ package.json
@@ -24,7 +24,8 @@
         "update": "nx migrate latest",
         "workspace-generator": "nx workspace-generator",
         "dep-graph": "nx dep-graph",
-        "help": "nx help"
+        "help": "nx help",
+        "tutorial": "jest tutorial/* -c tutorial/jest.config.ts"
     },
     "standard-version": {
         "bumpFiles": [
