+++ tutorial/tsconfig.json
@@ -1,0 +1,7 @@
+{
+    "extends": "../tsconfig.base.json",
+    "compilerOptions": {
+        "module": "commonjs",
+        "types": ["jest", "jest-extended", "node"]
+    }
+}
